refactor: rename Zettelkasten to ZettelStore, fix database imports
